Sort sequence list from XNAT
Before, the `seqlist` parameter would often give the right sequence, but on some datasets the sequence would be all wrong. XNAT does not sort the `/data/experiments/{experiment_id}/scan` results. Using the XNAT `sortBy` query string sorts in alpha order, not numeric order. I added the sort in the python code to compensate for the bad `sortBy` parameter, and added a test to verify the fix.