feat: add part-mag and -phase to BIDS names
A single scan exported by the scanner as separate magnitude and phase series is not yet appropriately given the part-mag and part-phase labels required by BIDS (https://github.com/nipy/heudiconv/issues/518). Instead, one is labeled as a duplicate. This fix adds a function to add the part entity to the series description, which heudiconv then treats correctly in the BIDS conversion.