feat: convert MRS DICOMS to BIDS format
Renames and copies DICOMS to bids/sourcedata, then converts to NIFTI format with json sidecars, and uses heudiconv functions to extract info for _scans.tsv file in the session directory and add a row for each new MRS NIFTI.

Also adds a function remove_func_acquisition_duration_field to bids_utils, because the addition of MRS data requires us to move to the new BIDS validator 2.0. This schema version of the validator enforces that functional data jsons may not have both the AcquisitionDuration and RepetitionTime keys. This function checks all func jsons and removes the AcquisitionDuration field if the RepetitionTime is specified.